+++ .github/actions/sync-labels/index.js
@@ -1,5 +1,6 @@
 import * as core from "@actions/core"; // https://github.com/actions/toolkit/tree/main/packages/core
 import * as github from "@actions/github"; // https://github.com/actions/toolkit/tree/main/packages/github
+import { getIssueNumber } from "../utils";
 
 /** 메인 액션 함수 */
 const run = async () => {
@@ -46,19 +47,6 @@ const getPRContext = (context) => {
   const repo = context.repo.repo;
 
   return { prNumber, branchName, owner, repo };
-};
-
-/**
- * 브랜치 이름에서 이슈 번호를 추출하는 함수
- * @param {string} branchName - 브랜치 이름
- * @returns {number} issueNumber - 이슈 번호
- */
-const getIssueNumber = (branchName) => {
-  const match = branchName.match(/#(\d+)/);
-  if (!match) return;
-
-  const issueNumber = parseInt(match[1], 10);
-  return issueNumber;
 };
 
 /**
