+++ .github/actions/utils/index.js
@@ -1,0 +1,1 @@
+export { default as getIssueNumber } from "./getIssueNumber";
